tried fixing Bo Li's method again, no dice
I realized the method is essentially just quantization to even or odd
numbers depending some bit comparisons. I tried rewriting with
simplified code in hopes that some of the error was being caused by
incorrect implementation in the more complex version.

But it still doesn't seem to work. I won't put more time into it,
because it's okay if it fails. We are ultimately using an algorithm that
was designed for image embedding in grayscale images for bit embedding
in color images