fixed a couple bugs with odd data sizes
LSB watermarking had two problems. The first was that the extractor did
not correctly extract an odd number of bytes. More specifically, when
the number of bits to extract was not a multiple of the number of bits
per pixel, it would leave off some bits and the message would be garbled
at the end.

Additionally, the test script used bin2ascii and os.urandom to generate
the arbitrarily long messages with "good" randomness. But it would fail
to generate an odd number of bytes because it relied on an integer
division by two. Now it just uses a dumb random.choice to create a
string.

Both of these errors cropped up when testing the Tapestry watermarking
application